+++ lib/src/enc_shared_pref_async.dart
@@ -36,12 +36,9 @@ class EncryptedSharedPreferencesAsync {
 
   /// Remove by checking condition.
   Future<bool> removeWhere(
-      {bool notify = true,
-      bool notifyEach = false,
-      required Function(String key) condition}) async {
+      {bool notify = true, required Function(String key) condition}) async {
     assert(_key != null);
-    return _decorator.removeWhere(
-        condition: condition, notify: notify, notifyEach: notifyEach);
+    return _decorator.removeWhere(condition: condition, notify: notify);
   }
 
   /// Remove the value associated with the specified key from SharedPreferences.
@@ -162,12 +159,6 @@ class EncryptedSharedPreferencesAsync {
   Stream<String> listenSet({required Set<String> keys}) {
     assert(_key != null);
     return _decorator.listenSet(keys: keys);
-  }
-
-  /// Notify listeners
-  Stream<String> notifyObservers() {
-    assert(_key != null);
-    return _decorator.notifyObservers();
   }
 
   ///Save map